+++ src/Data/Tuple/Append/Example.hs
@@ -1,12 +1,15 @@
-{-# LANGUAGE MagicHash, UnboxedTuples, TemplateHaskell #-}
+{-# LANGUAGE CPP, MagicHash, TemplateHaskell, UnboxedTuples #-}
 {-# OPTIONS_GHC -fobject-code #-}
 
 module Data.Tuple.Append.Example where
 
-import Data.Tuple.Append.TemplateHaskell(makeUnboxedTupleAppendFun)
+#if MIN_VERSION_ghc_prim(0,7,0)
+import Data.Tuple.Append.TemplateHaskell(makeBoxedTupleAppendFun, makeUnboxedTupleAppendFun)
 
-import GHC.Exts(Int#, Float#)
+import GHC.Exts(Float#, Int#)
 
-import Language.Haskell.TH.Syntax(Type(ConT), mkName)
+import Language.Haskell.TH.Syntax(Type(ConT, VarT), mkName)
 
-makeUnboxedTupleAppendFun (mkName "foo") [ ConT ''Int#, ConT ''Float# ] [ ConT ''Float#, ConT ''Int# ]
+makeBoxedTupleAppendFun (mkName "append_if_f") [ ConT ''Int, ConT ''Float ] [ ConT ''Float ]
+makeUnboxedTupleAppendFun (mkName "uappend_ix_f") [ ConT ''Int#, VarT (mkName "a")] [ConT ''Float# ]
+#endif