+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-# `tuple-append` Changelog
+# `tuple-append` changelog
 
-For a full list of changes, see the history on [GitHub](https://github.com/hapytex/tuple-append).
+For a full list of changes, see the history on [*GitHub*](https://github.com/hapytex/tuple-append).
+
+## Version 0.1.0.0
+
+A first intial version that allows constructing arbitrary prepend and append functions and typeclass instances. Instances for `AddL` and `AddR` are defined to construct tuples with a length up to 62, and it is possible to construct any tuple by appending tuples for a size up to 19. Template Haskell and quasiquoters are avaiable to construct functions to prepend and append tuples for both boxed and unboxed tuples.